+++ src/main/java/controller/UserController.java
@@ -8,9 +8,10 @@ import db.UserDatabase;
 import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import reader.RequestPostReader;
-import reader.RequestReader;
+import reader.requestReader.RequestPostReader;
+import reader.requestReader.RequestReader;
 import request.HttpRequest;
+import request.RequestDataType;
 import response.HttpResponse;
 import service.Service;
 import service.UserService;
