+++ src/main/java/reader/fileReader/TemplatesFileReader.java
@@ -2,9 +2,7 @@ package reader.fileReader;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import util.Url;
-import util.error.HttpsErrorMessage;
-import webserver.RequestHandler;
+import request.Url;
 
 import java.io.File;
 import java.io.IOException;
