+++ src/main/java/reader/requestReader/RequestReader.java
@@ -1,8 +1,7 @@
-package reader;
+package reader.requestReader;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import reader.fileReader.FileReader;
 import request.RequestHeader;
 import util.HttpMethod;
 import request.HttpRequest;
@@ -10,7 +9,6 @@ import util.error.HttpsErrorMessage;
 
 import java.net.ProtocolException;
 import java.util.HashMap;
-import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
