+++ src/main/java/request/RequestDataType.java
@@ -1,4 +1,4 @@
-package util;
+package request;
 
 import java.util.Arrays;
