+++ src/test/java/reader/fileReader/StaticFileReaderTest.java
@@ -2,8 +2,8 @@ package reader.fileReader;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import util.RequestDataType;
-import util.Url;
+import request.RequestDataType;
+import request.Url;
 
 import java.io.File;
 import java.io.IOException;
