+++ src/test/java/reader/fileReader/TemplatesFileReaderTest.java
@@ -3,8 +3,8 @@ package reader.fileReader;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import util.RequestDataType;
-import util.Url;
+import request.RequestDataType;
+import request.Url;
 
 import java.io.File;
 import java.io.IOException;
